# flulink/backend/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FluLink API",
    "version": "1.0.0",
    "description": "FluLink星尘共鸣版API文档",
    "contact": {
      "name": "FluLink Team",
      "email": "support@flulink.com",
      "url": "https://flulink.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "开发环境"
    },
    {
      "url": "https://flulink-app.zeabur.app",
      "description": "生产环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "错误信息"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/starseeds/publish": {
      "post": {
        "summary": "发布星种",
        "description": "用户发布新的星种内容",
        "tags": [
          "StarSeeds"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "星种内容",
                    "example": "分享一个有趣的技术发现"
                  },
                  "imageUrl": {
                    "type": "string",
                    "description": "图片URL",
                    "example": "https://example.com/image.jpg"
                  },
                  "audioUrl": {
                    "type": "string",
                    "description": "音频URL",
                    "example": "https://example.com/audio.mp3"
                  },
                  "spectrum": {
                    "type": "object",
                    "description": "光谱数据"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "星种发布成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "starseed": {
                          "$ref": "#/components/schemas/StarSeed"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未授权访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}